rename gnosisProtocol to cow
