+++ src/custom/api/cow/errors/OperatorError.ts
