+++ src/custom/api/cow/errors/QuoteError.ts
