+++ src/custom/api/cow/hooks.ts
@@ -1,7 +1,7 @@
 import useSWR from 'swr'
 
 import { useActiveWeb3React } from 'hooks/web3'
-import { getOrders, OrderMetaData } from 'api/gnosisProtocol/api'
+import { getOrders, OrderMetaData } from 'api/cow/api'
 import { AMOUNT_OF_ORDERS_TO_FETCH } from 'constants/index'
 
 export function useGpOrders(account?: string | null, refreshInterval?: number): OrderMetaData[] | undefined {
