+++ src/custom/components/Menu/index.tsx
@@ -17,7 +17,7 @@ import cowRunnerImage from 'assets/cow-swap/game.gif'
 import ninjaCowImage from 'assets/cow-swap/ninja-cow.png'
 import { ApplicationModal } from 'state/application/reducer'
 import { getExplorerAddressLink } from 'utils/explorer'
-import { useHasOrders } from 'api/gnosisProtocol/hooks'
+import { useHasOrders } from 'api/cow/hooks'
 import { useHistory } from 'react-router-dom'
 import CowClaimButton, { Wrapper as ClaimButtonWrapper } from 'components/CowClaimButton'
 import { IS_CLAIMING_ENABLED } from 'pages/Claim/const'
