+++ src/custom/hooks/useRefetchPriceCallback.tsx
@@ -1,12 +1,8 @@
 import { useCallback } from 'react'
 
 import { FeeQuoteParams, getBestQuote, getFastQuote, QuoteParams, QuoteResult } from 'utils/price'
-import { isValidOperatorError, ApiErrorCodes } from 'api/gnosisProtocol/errors/OperatorError'
-import GpQuoteError, {
-  GpQuoteErrorCodes,
-  GpQuoteErrorDetails,
-  isValidQuoteError,
-} from 'api/gnosisProtocol/errors/QuoteError'
+import { isValidOperatorError, ApiErrorCodes } from 'api/cow/errors/OperatorError'
+import GpQuoteError, { GpQuoteErrorCodes, GpQuoteErrorDetails, isValidQuoteError } from 'api/cow/errors/QuoteError'
 import { registerOnWindow, getPromiseFulfilledValue, isPromiseFulfilled } from 'utils/misc'
 
 import { isOnline } from 'hooks/useIsOnline'
