+++ src/custom/pages/Profile/index.tsx
@@ -28,7 +28,7 @@ import { MouseoverTooltipContent } from 'components/Tooltip'
 import NotificationBanner from 'components/NotificationBanner'
 import { SupportedChainId as ChainId } from 'constants/chains'
 import AffiliateStatusCheck from 'components/AffiliateStatusCheck'
-import { useHasOrders } from 'api/gnosisProtocol/hooks'
+import { useHasOrders } from 'api/cow/hooks'
 import { Title } from 'components/Page'
 import { useTokenBalance } from 'state/wallet/hooks'
 import { V_COW } from 'constants/tokens'
