+++ src/custom/state/gas/actions.ts
@@ -1,6 +1,6 @@
 import { createAction } from '@reduxjs/toolkit'
 import { WithChainId } from '../lists/actions'
-import { GasFeeEndpointResponse } from 'api/gnosisProtocol/api'
+import { GasFeeEndpointResponse } from 'api/cow/api'
 
 export type UpdateGasPrices = GasFeeEndpointResponse & WithChainId
