+++ src/custom/state/orders/helpers.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components/macro'
 
 import { formatOrderId, shortenOrderId } from 'utils'
-import { OrderID } from 'api/gnosisProtocol'
+import { OrderID } from 'api/cow'
 import { addPopup } from 'state/application/reducer'
 import { OrderStatus } from './actions'
 import { CancellationSummary } from 'components/AccountDetails/Transaction/styled'
