+++ src/custom/utils/price.ts
@@ -3,8 +3,8 @@ import BigNumberJs from 'bignumber.js'
 import * as Sentry from '@sentry/browser'
 import { Percent } from '@uniswap/sdk-core'
 
-import { getQuote, getPriceQuoteLegacy as getPriceQuoteGp, OrderMetaData } from 'api/gnosisProtocol'
-import GpQuoteError, { GpQuoteErrorCodes } from 'api/gnosisProtocol/errors/QuoteError'
+import { getQuote, getPriceQuoteLegacy as getPriceQuoteGp, OrderMetaData } from 'api/cow'
+import GpQuoteError, { GpQuoteErrorCodes } from 'api/cow/errors/QuoteError'
 import { getCanonicalMarket, isPromiseFulfilled, withTimeout } from 'utils/misc'
 import { formatAtoms } from 'utils/format'
 import { PRICE_API_TIMEOUT_MS } from 'constants/index'
